// src/chunithm/model/userCourse.ts
export interface UserCourseItem {
  courseId: number;
  classId: number;
  playCount: number;
  scoreMax: number;
  isFullCombo: boolean;
  isAllJustice: boolean;
  isSuccess: boolean;
  scoreRank: number;
  eventId: number;
  lastPlayDate: Date;
  param1: number;
  param2: number;
  param3: number;
  param4: number;
  isClear: boolean;
}